Update typespecs-and-behaviours.markdown
These changes attempt to address some of the mental friction around some of the concepts by smoothing out the readability, linking more relevant documents, and adding a few more examples.  For example, the explanation around https://hexdocs.pm/elixir/Kernel.html#round/1 was confusing because the sample code on this page did not match the code on the linked page.